feat: initial browser support
refactor: split out index into separate files for easier node and browser support
tests: make sure tests pass again
chore: use node: style imports
breaking: removed fs support on types/interface.ts
build: adjusted exports to work with node and node16 resolution modes
style: adjusted prettier quotes to match eslint
