+++ specs/invalid.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai'
 import { sync as globSync } from 'glob'
-import { imageSize } from '../lib'
+import { imageSize } from '../lib/node'
 
 // Test all invalid files
 describe('Invalid Images', () => {
